+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Don't add total number of files to `delete-orphaned-files` command output.
 - Added `id` field to Mastodon API `Application` object.
 - Improve validation of object IDs.
+- Validate actor ID and inbox URL before saving actor profile to database.
 
 ### Removed